refactor(service): correct package name
